refactor: move from metrics/models.py to data_models.py to include qa_generation models
