+++ app/src/evaluation/metrics/logging.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from typing import Any, Optional, TextIO
 
 from ..utils.jsonl_to_csv import convert_results_to_csv
-from .models import BatchConfig, EvaluationResult, MetricsSummary
+from src.evaluation.data_models import BatchConfig, EvaluationResult, MetricsSummary
 
 
 class EvaluationLogger:
